% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{arguments}
\alias{arguments}
\title{Extract arguments from an object}
\usage{
arguments(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{additional arguments (currently unused).}
}
\value{
A pairlist of formal arguments from the associated function,
  as returned by `formals()`.
}
\description{
This method extracts the formal arguments of an object.
For objects of `travelpath_spec` extracts the formal arguments of the fitting function
associated with a travelpath specification object.
}
\examples{
spec <- threshold_stops()
arguments(spec)
}
