% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{assert_spec}
\alias{assert_spec}
\title{Assert that a travelpath spec is valid}
\usage{
assert_spec(spec)
}
\arguments{
\item{spec}{an travelpath spec object.}
}
\value{
R `NULL` if valid, throws an error otherwise.
}
\description{
Assert that a travelpath spec is valid
}
