% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_cpt.R
\name{change_point_permtest}
\alias{change_point_permtest}
\title{Change Point Permutation Test Specification}
\usage{
change_point_permtest(
  mode = "segmentation",
  engine = "cpt",
  alpha = 0.05,
  output_col = "cp_id"
)
}
\arguments{
\item{mode}{The analysis mode. Only "segmentation" is supported.}

\item{engine}{The computational engine to use. Default is "cpt".}

\item{alpha}{Significance level for the permutation test (default: 0.05).}

\item{output_col}{Name of the output column added to the trackframe
(default: \code{"cp_id"}).}
}
\value{
A travelpath specification object for change point detection.
}
\description{
Constructs a travelpath specification for change point detection using a
  permutation test. This function creates a specification object for change
  point analysis using the `cpt::change_point_test` engine. The method
  detects change points in a sequence by comparing segments using a specified
  distance metric and assessing significance
via permutation testing.
}
\details{
This specification uses the `cpt::change_point_test` function to perform
change point detection via permutation testing.
}
\examples{
spec <- change_point_permtest(alpha = 0.01)

}
\seealso{
[cpt::change_point_test()]
}
