% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_smoove.R
\name{cvm_change_points}
\alias{cvm_change_points}
\title{Find Change Points}
\usage{
cvm_change_points(
  mode = "segmentation",
  engine = "smoove",
  windowsize = 1000,
  windowstep = 50,
  model = "UCVM",
  progress = TRUE,
  time_unit = "mins",
  .parallel = FALSE,
  clusterwidth = NULL,
  modelset = "all",
  spline = TRUE,
  criterion = "BIC",
  verbose = TRUE,
  output_col = "phase_id"
)
}
\arguments{
\item{mode}{a character vector of length one, for the type of model. The only
possible value for this model is "segmentation".}

\item{engine}{a character vector of length one, specifying what
computational engine to use for fitting. Possible engines are listed
below. The default for this model is `"smoove"`.}

\item{windowsize}{time window of analysis to scan, IMPORTANTLY: in units of time (T).}

\item{windowstep}{step (in time) by which the window advances. The smaller
the step, the slower but more thorough the estimation.}

\item{model}{model to fit for the change point sweep - typically the most
complex model in the candidate model set.}

\item{progress}{whether or not to show a progress bar}

\item{time_unit}{of the windowsize AND the windowstep. The default is
"hours" - can be any of "secs", "mins", "hours", "days", "weeks"
(See \code{\link[base]{difftime}}). Ignored if time is not POSIX.}

\item{.parallel}{if set TRUE, will use \code{\link[foreach]{foreach}} to
parallelize the optimization. Requires establishing the}

\item{clusterwidth}{A time span within which very close change points are
considered a single change point.}

\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM,
RACVM, or \code{all}, which includes all of them)}

\item{spline}{whether or not to use the spline approximation for the final estimate.}

\item{criterion}{selection criterion - either BIC or AIC (can be upper- or lowercased)}

\item{verbose}{a logical to control verbose output (default: \code{TRUE}).}

\item{output_col}{a character string giving the name of the output column
added to the trackframe (default: \code{"phase_id"}).}
}
\value{
A trackframe with an additional column indicating change points.
}
\description{
Find Change Points
}
\examples{
cvm_change_points()

}
