% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_engine.R
\name{cvm_estimate_phases}
\alias{cvm_estimate_phases}
\alias{cvm_phases}
\title{Smoove estimate phases}
\usage{
cvm_estimate_phases(
  data,
  criterion = "BIC",
  verbose = TRUE,
  phase_id_col = "phase_id"
)

cvm_phases(
  mode = "grouping",
  engine = "smoove",
  criterion = "BIC",
  verbose = TRUE,
  phase_id_col = "phase_id"
)
}
\arguments{
\item{data}{currently a trackframe}

\item{criterion}{an information criterion such as BIC, AIC etc.}

\item{verbose}{whether to print a summary of the results}

\item{phase_id_col}{a character vector of length one, specifying the name of the
column to be used for the phase identifiers. Default is "phase_id".}

\item{mode}{a character vector of length one, for the type of model. The only
possible value for this model is "grouping".}

\item{engine}{a character vector of length one, specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is `"smoove"`.}
}
\value{
An `infomap` specification.
}
\description{
Smoove estimate phases
}
\examples{
infomap()

# Specify the criterion
cvm_phases(
  criterion = "BIC"
)
}
