% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{diagnose_crw}
\alias{diagnose_crw}
\title{Autorcorrelation}
\usage{
diagnose_crw(data)
}
\arguments{
\item{data}{an object coercible to class \code{track_frame}}
}
\description{
applies acf
}
