% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{elapsed}
\alias{elapsed}
\title{Extract Elapsed Time}
\usage{
elapsed(x, ...)
}
\arguments{
\item{x}{object from which the elapsed time is extracted.}

\item{\ldots}{further arguments currently ignored.}
}
\value{
an object of class \code{"proc_time"} as returned by
  \code{\link[base]{proc.time}}
}
\description{
The timing information of a given \code{"travelpath_fit"} can
be accessed via the method \code{'elapsed'}.
}
