% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{estimate_ucvm}
\alias{estimate_ucvm}
\title{Title}
\usage{
estimate_ucvm(
  tf,
  time.units = "min",
  diagnose = TRUE,
  CI = TRUE,
  method = "crw",
  ...
)
}
\arguments{
\item{tf}{}

\item{time.units}{}

\item{diagnose}{}

\item{CI}{}

\item{method}{}

\item{...}{}
}
\description{
Title
}
