% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_engine.R
\name{find_candidate_change_points}
\alias{find_candidate_change_points}
\title{Find Candidate Change Points}
\usage{
find_candidate_change_points(windowsweep, clusterwidth = 0, verbose = TRUE)
}
\arguments{
\item{windowsweep}{A windowsweep object (matrix), output of \code{\link{sweepRACVM}} function}

\item{clusterwidth}{A time span within which very close change points are considered a single chagne point.}

\item{verbose}{Whether or not to report the number of change points that are clustered away.}
}
\description{
A slightly modified version of smoove::findCandidateChangePoints. Also stores the minimum
clusterwidth in the attributes of the output.
}
\details{
The raw output of the window sweep (\code{\link{sweepRACVM}})
}
\examples{
library(smoove)
library(travelpaths)
data(simSweep,package = "smoove")
# The warning lets us know that some of the candidate change points
# are rather too close (in time) to each other
cp_all <- travelpaths:::find_candidate_change_points(windowsweep = simSweep, clusterwidth = 0)
as.vector(cp_all)
cp_cl <- travelpaths:::find_candidate_change_points(windowsweep = simSweep, clusterwidth = 4)
as.vector(cp_cl)
}
