% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{find_single_break_point}
\alias{find_single_break_point}
\alias{find_single_break_point.trackframe}
\alias{find_single_break_point.data.frame}
\alias{find_single_break_point.move2}
\alias{find_single_break_point.sftrack}
\title{Find single change point}
\usage{
find_single_break_point(
  data,
  k = 1,
  method = "sweep",
  plotme = TRUE,
  time.units = "min",
  ...
)

\method{find_single_break_point}{trackframe}(
  data,
  k = 1,
  method = "sweep",
  plotme = TRUE,
  time.units = "min",
  ...
)

\method{find_single_break_point}{data.frame}(
  data,
  k = 1,
  method = "sweep",
  plotme = TRUE,
  time.units = "min",
  ...
)

\method{find_single_break_point}{move2}(
  data,
  k = 1,
  method = "sweep",
  plotme = TRUE,
  time.units = "min",
  ...
)

\method{find_single_break_point}{sftrack}(
  data,
  k = 1,
  method = "sweep",
  plotme = TRUE,
  time.units = "min",
  ...
)
}
\arguments{
\item{data}{an object of class \code{trackframe}}

\item{k}{tuning parameter for the smoothing of the likelihood profile spline. 
The number of knots is "length(T)/4 * k" - the lower the value of k, the smoother the spline.}

\item{method}{one of "sweep" or "optimize". See details.}

\item{plotme}{whether to plot the resulting likelihood (only if method is "sweep").}

\item{time.units}{time units of calculations (e.g. "sec", "min", "hour", "day")}

\item{...}{additional parameters to pass to \code{\link{estimateUCVM}} function, in particular the method of estimation.  Under most conditions, fairly reliable and fast results are provided by the default \code{vLike} (velocity likelihood) method.}
}
\description{
Finds a single change point in UCVM parameters (time scale \eqn{tau} and rms 
speed \eqn{eta}) of a movement.
}
\details{
Two methods are provided: "sweep", which scans a set of possible 
change points, smooths the likelihoods and selects the maximum, or "optimize" 
which uses R's single dimension optimization algorithms to find the most 
likely change point. The latter is faster, but can be unreliable because the 
likelihood profiles are typically quite rough.
}
\examples{
set.seed(2025)
tf <- sim_travel_path(50, format = "trackframe")
find_single_break_point(tf, method = "sweep")
find_single_break_point(tf, method = "optimize")
}
