% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{n_segments}
\alias{n_segments}
\alias{n_groups}
\title{Accessor Functions for Segments and Groups}
\usage{
n_segments(x, ...)

n_groups(x, ...)
}
\arguments{
\item{x}{object from which the number of groups or segments are extracted.}

\item{\ldots}{further arguments currently ignored.}
}
\value{
an integer vector giving the number of segments or groups for each track.
}
\description{
The number of segments or groups produced by a \code{"travelpath_fit"}
can be accessed via \code{'n_segments'} and \code{'n_groups'}.
These are mode-specific:
  use \code{n_segments()} for a \code{"segmentation"} fit,
  and \code{n_groups()} for a \code{"grouping"} fit.
}
