% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{new_travelpath_fit}
\alias{new_travelpath_fit}
\title{Travelpath Fit}
\usage{
new_travelpath_fit(spec, fit, elapsed = NULL, preproc = list(), meta = NULL)
}
\arguments{
\item{spec}{an object of class \code{"travelpath_spec"},
giving the original specification storing the options used to fit.}

\item{fit}{the engine output, stored as-is.}

\item{elapsed}{an object of class \code{"proc_time"} (as returned by
\code{\link[base]{proc.time}}) giving how long the fit took.}

\item{preproc}{a list of preprocessing artifacts, currently a
placeholder (empty list) reserved for future predict
data transformations.}

\item{meta}{an optional list for storing additional information.}
}
\value{
an object of class
  \code{c("<engine>_fit", "<mode>_fit", "travelpath_fit")}, where
  \code{<mode>} is either \code{"segmentation"} or \code{"grouping"}
  taken from \code{spec[["mode"]]}.
}
\description{
Constructor for the fit \code{"travelpath_fit"} object.
}
