% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{new_travelpath_spec}
\alias{new_travelpath_spec}
\title{Constructor for travelpath specification objects}
\usage{
new_travelpath_spec(cls, args, eng_args, mode, engine)
}
\arguments{
\item{cls}{a character string for the class of the specification (e.g., "stop_detection").}

\item{args}{a list of main arguments for the specification.}

\item{eng_args}{a list of engine-specific arguments, will be typically an empty `list()` or
`NULL` when creating a new specification.}

\item{mode}{a character string for the type of travel path analysis.
Possible values are "segmentation" or "grouping" or "unknown".}

\item{engine}{a character string for the computational engine.}
}
\value{
A travelpath specification object with class `cls`.
}
\description{
`new_travelpath_spec()` is a developer-oriented constructor function for
creating new travelpath specification objects. This function is used
internally by user-facing specification functions.
}
