% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{plot_phases}
\alias{plot_phases}
\title{Plot Phases}
\usage{
plot_phases(
  x,
  cols = NULL,
  plot.parameters = TRUE,
  parameters = NULL,
  plot_legend = TRUE,
  legend_where = "bottomright",
  layout = c("horizontal", "vertical")
)
}
\arguments{
\item{x}{an object of class trackframe containing a cvm phases object in
the attributes or a phase list}

\item{cols}{colors for phases (by default uses the `rich.colors` palette
from `gplots`)}

\item{plot.parameters}{whether the parameters are plotted}

\item{parameters}{which parameters to plot (by default - ALL of the
estimated parameters)}

\item{plot_legend}{whether to plot a legend}

\item{legend_where}{location of legend}

\item{layout}{"horizontal" (default) or "vertical" - as preferred (partial
string matching accepted)}
}
\description{
Plot Phases
}
