% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{register_argument_mapping}
\alias{register_argument_mapping}
\title{Register an argument mapping for a model specification}
\usage{
register_argument_mapping(model, mode, engine, argument_mapping)
}
\arguments{
\item{model}{a character string giving the model name, e.g., `"threshold_stops"`.}

\item{mode}{a character string giving the mode (currently `"segmentation"`
and `"grouping"` are supported).}

\item{engine}{a character string giving the engine, e.g., `"smoove"`.}

\item{argument_mapping}{a named character vector mapping the names used in
`object[["args"]]` to the actual argument names expected by the fit
function. The format is `c("old_name_1" = "new_name_1",
"old_name_2" = "new_name_2", ...)`.}
}
\value{
Invisible `NULL`.
}
\description{
Register an argument mapping for a model specification
}
