% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{register_fit_function}
\alias{register_fit_function}
\title{Register a new fit function}
\usage{
register_fit_function(model, mode, engine, fit_function)
}
\arguments{
\item{model}{a character string giving the model name, e.g.,
`"threshold_stops"`.}

\item{mode}{a character string giving the mode, currently `"segmentation"`
and `"grouping"` are supported.}

\item{engine}{a character string giving the engine, e.g., `"smoove"`.}

\item{fit_function}{a function to be used for fitting.}
}
\value{
Invisible `NULL`.
}
\description{
Register a new fit function
}
\examples{
\dontrun{
register_fit_function("threshold_stops", "segmentation", "infostop",
  infostop::identify_stops)
}

}
