% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{segment_col}
\alias{segment_col}
\alias{group_col}
\title{Accessor Functions to Obtain Segment / Group Column}
\usage{
segment_col(x, ...)

group_col(x, ...)
}
\arguments{
\item{x}{the object from which the column name should be extracted.}

\item{\ldots}{additional arguments currently ignored.}
}
\value{
a character vector giving the name of the column where the
  output is stored.
}
\description{
The name of the column in the fit where the newly created
segmentation or grouping is stored.
}
