% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_trackframe.R
\name{sim_travel_paths}
\alias{sim_travel_paths}
\title{Generate Multiple Random Travel Paths}
\usage{
sim_travel_paths(
  ntracks,
  sizes,
  max_step = 0.001,
  time_increment = 60,
  start_location = c(48.2083537, 16.3725042),
  start_time = Sys.time(),
  stay_prob = 0.2,
  track_prefix = "track",
  format = c("trackframe", "data.frame", "matrix", "sftrack", "move2")
)
}
\arguments{
\item{ntracks}{An integer specifying the number of tracks to generate.}

\item{sizes}{An integer vector specifying the number of points for each track. If a single value
is provided, it will be repeated for all tracks.}

\item{max_step}{A numeric value specifying the maximum step size for random movements.
Default is 0.001 (approximately 100m at the equator).}

\item{time_increment}{A numeric giving the time between consecutive points in seconds.
Default is 60 (1 minute).}

\item{start_location}{A numeric vector giving the starting location as northing, easting.}

\item{start_time}{A POSIXct giving the starting time for the paths. Default is current time.}

\item{stay_prob}{A numeric between 0 and 1 giving the probability of staying at the same location.
Default is 0.2.}

\item{track_prefix}{A character string used as a prefix for track IDs. Default is "track".
Track IDs will be formatted as \code{prefix\_number}.}

\item{format}{A character string, either \code{"trackframe"} (easting/northing), \code{"data.frame"}, \code{"matrix"}, \code{"sftrack"} or \code{"move2"}.}
}
\value{
Depending on the format argument either a \code{"trackframe"} or
 \code{"data.frame"} or \code{"matrix"} or \code{"sftrack"} or \code{"move2"}.
}
\description{
This function creates multiple random travel paths and combines them into a single object. 
Each path is assigned coordinates (easting/northing for format trackframe and latitude/longitude else) and time values and a unique track ID.
Each path can include stationary periods and movements with configurable parameters. The following output formats are supported: \code{"trackframe"} (easting/northing), \code{"data.frame"}, \code{"matrix"}, \code{"sftrack"} or \code{"move2"}.
}
\examples{
# Generate 3 tracks with different sizes
ntracks <- 3
sizes <- c(2, 4, 5)
multi_track <- sim_travel_paths(ntracks, sizes)

# Generate 5 tracks all with the same size
uniform_tracks <- sim_travel_paths(5, 10)

# Extract a specific track
track2 <- select_id(multi_track, "track_2")
}
