% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{specification}
\alias{specification}
\title{Specification Accessor Function}
\usage{
specification(x, ...)
}
\arguments{
\item{x}{object from which the specification is extracted.}

\item{\ldots}{further arguments currently ignored.}
}
\value{
an object inheriting from \code{"travelpath_spec"}
}
\description{
The \link{specification} of a given travelpath fit
(\code{"travelpath_fit"}) can be accessed via the method \code{'specification'}.
}
