% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{summarize_phases}
\alias{summarize_phases}
\title{Summarize phases}
\usage{
summarize_phases(x)
}
\arguments{
\item{x}{an object of class trackframe containing a cvm phases object in the attributes or a
phase list}
}
\description{
Takes a phase list and obtains a tidy table of point estimates of relevant parameters,
}
