% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{sweep_racvm}
\alias{sweep_racvm}
\title{Title}
\usage{
sweep_racvm(
  tf,
  windowsize = 1000,
  windowstep = 50,
  model = "UCVM",
  time.unit = "mins",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{tf}{}

\item{windowsize}{}

\item{windowstep}{}

\item{model}{}

\item{time.unit}{}

\item{.parallel}{}

\item{...}{}
}
\description{
Title
}
