% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{test_cp}
\alias{test_cp}
\alias{test_cp.trackframe}
\alias{test_cp.data.frame}
\alias{test_cp.move2}
\alias{test_cp.sftrack}
\title{Test RACVM change point}
\usage{
test_cp(
  data,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  time.units = "min",
  ...
)

\method{test_cp}{trackframe}(
  data,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  time.units = "min",
  ...
)

\method{test_cp}{data.frame}(
  data,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  time.units = "min",
  ...
)

\method{test_cp}{move2}(
  data,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  time.units = "min",
  ...
)

\method{test_cp}{sftrack}(
  data,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  time.units = "min",
  ...
)
}
\arguments{
\item{data}{an object of class \code{trackframe}}

\item{cp}{change point}

\item{start}{beginning time of segment to analyze}

\item{end}{end time of segment to analyze}

\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM, RACVM, or \code{all}, which includes all of them)}

\item{spline}{whether or not to use the spline approximation for the final estimate.}

\item{criterion}{selection criterion - either BIC or AIC (can be upper- or lowercased)}

\item{time.units}{time units of calculations (e.g. "sec", "min", "hour", "day")}

\item{...}{further params to \code{getFit} internal function}
}
\description{
Identifies appropriate models on either side of a change point
}
\examples{
tf <- sim_travel_path(1000, format = "trackframe")
test_cp(tf, 10, 1, 100)
}
