% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infostop.R
\name{threshold_stops}
\alias{threshold_stops}
\title{Threshold-based Stop Detection}
\usage{
threshold_stops(
  mode = "segmentation",
  engine = "infostop",
  r1 = 10,
  min_size = 2L,
  min_staying_time = 300L,
  max_time_between = 86400L,
  output_col = "stop_id"
)
}
\arguments{
\item{mode}{a character vector of length one, for the type of model. The only
possible value for this model is "stops".}

\item{engine}{a character vector of length one, specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is `"infostop"`.}

\item{r1}{A positive numeric value giving the maximum distance between
time-consecutive points to label them as stationary. Higher values will
result in more points being considered stationary.}

\item{min_size}{A positive integer giving the minimum number of points
required to consider a group stationary.}

\item{min_staying_time}{A positive integer giving the minimum duration
(in seconds) that can constitute a stop. Only relevant if timestamps
are provided in the data.}

\item{max_time_between}{A positive integer giving the maximum duration
(in seconds) between consecutive points to consider them part of the
same stop. Only relevant if timestamps are provided.}

\item{output_col}{a character vector of length one, specifying the name of the
column to be used for the stop identifiers. Default is "stop_id".}
}
\value{
A `threshold_stops` specification.
}
\description{
`threshold_stops()` defines a model that identifies stops in movement data
based on distance and time thresholds. This method detects stationary
periods by grouping consecutive points that are within a specified distance
and time criteria.

There are different ways to fit this model, and the method of estimation is
chosen by setting the model engine. The engine-specific pages for this model
are listed below.

- \code{infostop::identify_stops}: Threshold-based stop detection
}
\details{
## What does it mean to predict?

For a threshold-based stop detection model, stops are identified by grouping
consecutive points that meet the distance and time criteria. Therefore,
prediction in travelpaths assigns stop identifiers to movement trajectory
points based on these threshold conditions.
}
\examples{
threshold_stops()

# Specify distance metric and parameters
ths_spec <- threshold_stops(
  r1 = 20,
  min_size = 3
)
}
