% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_path_matrix}
\alias{travel_path_matrix}
\title{Simulated Travel Path}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 4 columns.
}
\usage{
travel_path_matrix
}
\description{
matrix created with sim_travel_path().
}
\keyword{datasets}
