% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_path_sftrack}
\alias{travel_path_sftrack}
\title{Simulated Travel Path}
\format{
An object of class \code{sftrack} (inherits from \code{sf}, \code{data.frame}) with 1000 rows and 6 columns.
}
\usage{
travel_path_sftrack
}
\description{
sftrack object created with sim_travel_path().
}
\keyword{datasets}
