% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_path_trackframe}
\alias{travel_path_trackframe}
\title{Simulated Travel Path}
\format{
An object of class \code{trackframe} (inherits from \code{data.frame}) with 1000 rows and 6 columns.
}
\usage{
travel_path_trackframe
}
\description{
track_frame created with sim_travel_path().
}
\keyword{datasets}
