% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_paths_data_frame}
\alias{travel_paths_data_frame}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{data.frame} with 3000 rows and 4 columns.
}
\usage{
travel_paths_data_frame
}
\description{
data.frame created with sim_travel_paths().
}
\keyword{datasets}
