% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_paths_sftrack}
\alias{travel_paths_sftrack}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{sftrack} (inherits from \code{sf}, \code{data.frame}) with 3000 rows and 6 columns.
}
\usage{
travel_paths_sftrack
}
\description{
sftrack object created with sim_travel_paths().
}
\keyword{datasets}
