% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{travel_paths_trackframe}
\alias{travel_paths_trackframe}
\title{Simulated Travel Path with multiple IDs}
\format{
An object of class \code{trackframe} (inherits from \code{data.frame}) with 3000 rows and 6 columns.
}
\usage{
travel_paths_trackframe
}
\description{
trackframe created with sim_travel_paths().
}
\keyword{datasets}
