% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{diagnose_crw}
\alias{diagnose_crw}
\alias{diagnose_crw.trackframe}
\alias{diagnose_crw.data.frame}
\alias{diagnose_crw.move2}
\alias{diagnose_crw.sftrack}
\title{Autorcorrelation}
\usage{
diagnose_crw(data)

\method{diagnose_crw}{trackframe}(data)

\method{diagnose_crw}{data.frame}(data)

\method{diagnose_crw}{move2}(data)

\method{diagnose_crw}{sftrack}(data)
}
\arguments{
\item{data}{an object coercible to class \code{trackframe}}
}
\description{
applies acf
}
