% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{estimate_racvm}
\alias{estimate_racvm}
\title{Title}
\usage{
estimate_racvm(tf, time.units = "min", ...)
}
\arguments{
\item{tf}{}

\item{time.units}{}

\item{...}{}
}
\description{
Title
}
