% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{estimate_ucvm}
\alias{estimate_ucvm}
\alias{estimate_ucvm.trackframe}
\alias{estimate_ucvm.data.frame}
\alias{estimate_ucvm.move2}
\alias{estimate_ucvm.sftrack}
\title{Estimating parameters of unbiased CVM}
\usage{
estimate_ucvm(
  data,
  method = c("vaf", "crw", "vLike", "zLike", "crawl")[3],
  parameters = c("tau", "nu"),
  time_units = "mins",
  ci = FALSE,
  spline = FALSE,
  diagnose = TRUE,
  ...
)

\method{estimate_ucvm}{trackframe}(
  data,
  method = c("vaf", "crw", "vLike", "zLike", "crawl")[3],
  parameters = c("tau", "nu"),
  time_units = "mins",
  ci = FALSE,
  spline = FALSE,
  diagnose = TRUE,
  ...
)

\method{estimate_ucvm}{data.frame}(
  data,
  method = c("vaf", "crw", "vLike", "zLike", "crawl")[3],
  parameters = c("tau", "nu"),
  time_units = "mins",
  ci = FALSE,
  spline = FALSE,
  diagnose = TRUE,
  ...
)

\method{estimate_ucvm}{move2}(
  data,
  method = c("vaf", "crw", "vLike", "zLike", "crawl")[3],
  parameters = c("tau", "nu"),
  time_units = "mins",
  ci = FALSE,
  spline = FALSE,
  diagnose = TRUE,
  ...
)

\method{estimate_ucvm}{sftrack}(
  data,
  method = c("vaf", "crw", "vLike", "zLike", "crawl")[3],
  parameters = c("tau", "nu"),
  time_units = "mins",
  ci = FALSE,
  spline = FALSE,
  diagnose = TRUE,
  ...
)
}
\arguments{
\item{data}{an object coercible to class \code{trackframe}}

\item{method}{the method to use for the estimation. These are (in
increasing): velocity auto-correlation fitting (\code{vaf}), correlated
random walk matching (\code{crw}), velocity likelihood (\code{vLike}),
position likelihood (\code{zLike}) and position likelihood with Kalman
filter (\code{crawl}). This last method is generally the best method,
since it fits the position likelihood more efficiently by using a Kalman
filter. It is based on Johnson et al (2008) and is a wrapper for the
\code{\link[crawl]{crwMLE}} in the \code{\link[crawl]{crawl}} package.
The default method is \code{vLike}.}

\item{parameters}{which parameters to estimate. For most methods "tau" and
"nu" are always both estimated, but some computation can be saved for the
velocity likelihood method by providing an estimate for "nu".}

\item{time_units}{unit of time to be passed to
\code{\link[base]{difftime}()} to compute a relative numeric time vector.}

\item{ci}{whether or not to compute 95\% confidence intervals for
parameters. In some cases, this can slow the computation down somewhat.}

\item{spline}{whether or not to use the spline correction (only relevant for
\code{vaf} and \code{vLike}).}

\item{diagnose}{whether to draw a diagnostic plot. Varies for different
methods.}

\item{...}{additional parameters to pass to estimation functions. These are
particularly involved in the \code{crawl} method (see
\code{\link[crawl]{crwMLE}}).}
}
\value{
A data frame with point estimates of mean speed `nu` and time-scale
  `tau`
}
\description{
This function estimates the mean speed \eqn{nu}, the time-scale \eqn{tau}
  and (occasionally) the initial speed \eqn{v_0} of the unbiased correlated
  velocity movement (UCVM). See Gurarie et al. (2016) and the
  \code{vignette("smoove", package = "smoove")} vignette for more details.
}
