% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoove_wrapper_functions.R
\name{test_cp}
\alias{test_cp}
\title{Test RACVM change point}
\usage{
test_cp(
  tf,
  cp,
  start,
  end,
  modelset = "all",
  spline = FALSE,
  criterion = "BIC",
  ...
)
}
\arguments{
\item{tf}{an object of class \code{track_frame}}

\item{cp}{change point}

\item{start}{beginning time of segment to analyze}

\item{end}{end time of segment to analyze}

\item{modelset}{set of models to compare (combination of UCVM, ACVM, RCVM, RACVM, or \code{all}, which includes all of them)}

\item{spline}{whether or not to use the spline approximation for the final estimate.}

\item{criterion}{selection criterion - either BIC or AIC (can be upper- or lowercased)}

\item{...}{further params to \code{getFit} internal function}
}
\description{
Identifies appropriate models on either side of a change point
}
